Reorganization and bug fixes
