+++ assets/js/editor.js
@@ -1,1 +1,1 @@
-this.TeamList=function(n){var r={};function o(e){if(r[e])return r[e].exports;var t=r[e]={i:e,l:!1,exports:{}};return n[e].call(t.exports,t,t.exports,o),t.l=!0,t.exports}return o.m=n,o.c=r,o.d=function(e,t,n){o.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.t=function(t,e){if(1&e&&(t=o(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(o.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)o.d(n,r,function(e){return t[e]}.bind(null,r));return n},o.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(t,"a",t),t},o.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},o.p="",o(o.s=19)}([function(e,t){!function(){e.exports=this.wp.element}()},function(e,t){!function(){e.exports=this.wp.i18n}()},function(e,t){!function(){e.exports=this.wp.components}()},function(e,t){!function(){e.exports=this.wp.data}()},function(e,t,n){var o=n(5);e.exports=function(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(e){o(t,e,n[e])})}return t}},function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t){e.exports=function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}},function(e,t){function r(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}e.exports=function(e,t,n){return t&&r(e.prototype,t),n&&r(e,n),e}},function(e,t,n){var r=n(16),o=n(17);e.exports=function(e,t){return!t||"object"!==r(t)&&"function"!=typeof t?o(e):t}},function(t,e){function n(e){return t.exports=n=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)},n(e)}t.exports=n},function(e,t,n){var r=n(18);e.exports=function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&r(e,t)}},function(e,t){!function(){e.exports=this.wp.blocks}()},function(e,t){!function(){e.exports=this.wp.editor}()},function(e,t){function a(e,t,n,r,o,i,s){try{var c=e[i](s),a=c.value}catch(e){return void n(e)}c.done?t(a):Promise.resolve(a).then(r,o)}e.exports=function(c){return function(){var e=this,s=arguments;return new Promise(function(t,n){var r=c.apply(e,s);function o(e){a(r,t,n,o,i,"next",e)}function i(e){a(r,t,n,o,i,"throw",e)}o(void 0)})}}},function(e,t){!function(){e.exports=this.wp.apiFetch}()},function(e,t,n){},function(t,e){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function r(e){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?t.exports=r=function(e){return n(e)}:t.exports=r=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},r(e)}t.exports=r},function(e,t){e.exports=function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}},function(n,e){function r(e,t){return n.exports=r=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e},r(e,t)}n.exports=r},function(e,t,n){"use strict";n.r(t);var r,f=n(0),m=n(1),o=n(11),l=n(12),p=n(2),i=(n(15),n(6)),s=n.n(i),c=n(7),a=n.n(c),u=n(8),b=n.n(u),d=n(9),y=n.n(d),w=n(10),h=n.n(w),O=n(13),j=n.n(O),_=n(5),v=n.n(_),g=n(4),S=n.n(g),E=n(3),x=n(14),L=n.n(x),k={users:{},isLoading:{}};Object(E.registerStore)("wp-team-list",{reducer:function(){var e=0<arguments.length&&void 0!==arguments[0]?arguments[0]:k,t=1<arguments.length?arguments[1]:void 0;switch(t.type){case"SET_USERS":return S()({},e,{users:S()({},e.users,v()({},t.query,t.users))});case"SET_IS_LOADING":return S()({},e,{isLoading:S()({},e.isLoading,v()({},t.query,t.isLoading))})}return e},actions:{setUsers:function(e,t){return{type:"SET_USERS",query:e,users:t}},setLoading:function(e,t){return{type:"SET_IS_LOADING",query:e,isLoading:t}}},selectors:{getUsers:function(e,t){return e.users[t]},isLoading:function(e,t){var n=e.isLoading;return n[t]}},resolvers:{getUsers:(r=j()(regeneratorRuntime.mark(function e(t,n){var r;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return Object(E.dispatch)("wp-team-list").setLoading(n,!0),e.next=3,L()({path:"/wp-team-list/v1/users?"+n});case 3:r=e.sent,Object(E.dispatch)("wp-team-list").setLoading(n,!1),Object(E.dispatch)("wp-team-list").setUsers(n,r);case 6:case"end":return e.stop()}},e,this)})),function(e,t){return r.apply(this,arguments)})}});var C=function(e){function t(){return s()(this,t),b()(this,y()(t).apply(this,arguments))}return h()(t,e),a()(t,[{key:"render",value:function(){var e=this.props,t=e.user,n=t.id,r=t.role,o=t.role_display_name,i=t.display_name,s=t.avatar_urls,c=t.description,a=t.link,u=t.post_count,l=e.showDescription,p=e.className;return Object(f.createElement)("div",{className:"wp-team-member wp-team-list-item author-".concat(n," role-").concat(r," ").concat(p)},Object(f.createElement)("figure",{className:"wp-team-member-avatar author-image"},Object(f.createElement)("img",{src:s[48],srcSet:"".concat(s[96]," 2x"),alt:"",width:"90"})),Object(f.createElement)("h2",{className:"wp-team-member-name"},i),o&&Object(f.createElement)("p",{className:"wp-team-member-role"},o),l&&Object(f.createElement)("p",{className:"wp-team-member-description"},c),u&&Object(f.createElement)("p",{className:"wp-team-member-posts-link"},Object(f.createElement)("a",{href:a,title:Object(m.sprintf)(Object(m.__)("View all posts by %s","wp-team-list"),i)},Object(m.sprintf)(Object(m._n)("View %s post","View %s posts",u,"wp-team-list"),u))))}}]),t}(f.Component),P=function(e){function t(){return s()(this,t),b()(this,y()(t).apply(this,arguments))}return h()(t,e),a()(t,[{key:"render",value:function(){var e=this.props,t=e.users,n=e.showLink,r=e.showDescription,o=e.isLoading;if(o&&!t)return Object(f.createElement)(p.Spinner,null);if(!t)return Object(f.createElement)("p",null,Object(m.__)("There are no users to show.","wp-team-list"));var i=t.map(function(e){return Object(f.createElement)(C,{user:e,showDescription:r,key:e.id})});return Object(f.createElement)(f.Fragment,null,o&&Object(f.createElement)(p.Spinner,null),i,n&&Object(f.createElement)("a",{href:"",className:"show-all"},Object(m.__)("Show all team members","wp-team-list")))}}]),t}(f.Component),N=Object(E.withSelect)(function(e,t){var n=t.number,r=t.roles,o=t.orderBy,i=t.order,s=t.showLink,c=t.showDescription,a=e("wp-team-list"),u=a.getUsers,l=a.isLoading,p={order:i,order_by:o,roles:r,per_page:n},f=Object.keys(p).map(function(e){return encodeURIComponent(e)+"="+encodeURIComponent(p[e])}).reduce(function(e,t){return"".concat(e,"&").concat(t)});return{users:u(f),isLoading:l(f),showLink:s,showDescription:c}})(P);Object(o.registerBlockType)("required/wp-team-list",{title:Object(m.__)("Team List","wp-team-list"),description:Object(m.__)("Display website authors.","wp-team-list"),category:"widgets",icon:"groups",keywords:[Object(m.__)("team","wp-team-list"),Object(m.__)("list","wp-team-list"),Object(m.__)("authors","wp-team-list")],supports:{anchor:!1,customClassName:!1,html:!1},attributes:{number:{type:"number",default:10},showLink:{type:"boolean",default:!0},showDescription:{type:"boolean",default:!0},roles:{type:"array",default:["administrator"]},orderBy:{type:"string",default:"post_count"},order:{type:"string",default:"desc"}},edit:function(e){var t=e.className,n=e.attributes,r=n.number,o=n.showLink,i=n.showDescription,s=n.roles,c=n.orderBy,a=n.order,u=e.setAttributes;return Object(f.createElement)(f.Fragment,null,Object(f.createElement)(N,{number:r,showLink:o,showDescription:i,roles:s,orderBy:c,order:a,className:t}),Object(f.createElement)(l.InspectorControls,null,Object(f.createElement)(p.PanelBody,null,Object(f.createElement)(p.TextControl,{label:Object(m.__)("Number of users to display","wp-team-list"),type:"number",value:r,onChange:function(e){u({number:e})}}),Object(f.createElement)(p.SelectControl,{label:Object(m.__)("Roles","wp-team-list"),help:Object(m.__)("Only show users with the selected roles","wp-team-list"),value:s,options:[{value:"administrator",label:"Administrator"},{value:"subscriber",label:"Subscriber"},{value:"editor",label:"Editor"}],multiple:!0,onChange:function(e){u({roles:e})}}),Object(f.createElement)(p.SelectControl,{label:Object(m.__)("Order By","wp-team-list"),value:c,options:[{value:"post_count",label:Object(m.__)("Post Count","wp-team-list")},{value:"first_name",label:Object(m.__)("First Name","wp-team-list")},{value:"last_name",label:Object(m.__)("Last Name","wp-team-list")}],onChange:function(e){u({orderBy:e})}}),Object(f.createElement)(p.SelectControl,{label:Object(m.__)("Order","wp-team-list"),value:a,options:[{value:"asc",label:Object(m.__)("Ascending","wp-team-list")},{value:"desc",label:Object(m.__)("Descending","wp-team-list")}],onChange:function(e){u({order:e})}}),Object(f.createElement)(p.ToggleControl,{label:Object(m.__)("Show user description","wp-team-list"),checked:i,onChange:function(e){u({showDescription:e})}}))))},save:function(){return null}})}]);
+this.TeamList=function(n){var r={};function o(e){if(r[e])return r[e].exports;var t=r[e]={i:e,l:!1,exports:{}};return n[e].call(t.exports,t,t.exports,o),t.l=!0,t.exports}return o.m=n,o.c=r,o.d=function(e,t,n){o.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.t=function(t,e){if(1&e&&(t=o(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(o.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)o.d(n,r,function(e){return t[e]}.bind(null,r));return n},o.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(t,"a",t),t},o.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},o.p="",o(o.s=29)}([function(e,t){!function(){e.exports=this.wp.element}()},function(e,t){!function(){e.exports=this.wp.i18n}()},function(e,t){!function(){e.exports=this.wp.data}()},function(e,t){!function(){e.exports=this.wp.components}()},function(e,t){e.exports=function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}},function(e,t){function r(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}e.exports=function(e,t,n){return t&&r(e.prototype,t),n&&r(e,n),e}},function(e,t,n){var r=n(21),o=n(12);e.exports=function(e,t){return!t||"object"!==r(t)&&"function"!=typeof t?o(e):t}},function(t,e){function n(e){return t.exports=n=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)},n(e)}t.exports=n},function(e,t,n){var r=n(22);e.exports=function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&r(e,t)}},function(e,t,n){var o=n(11);e.exports=function(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(e){o(t,e,n[e])})}return t}},function(e,t){!function(){e.exports=this.wp.apiFetch}()},function(e,t){e.exports=function(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}},function(e,t){e.exports=function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}},function(e,t,n){"use strict";var r=n(23),o=n(24),i=n(16);e.exports={formats:i,parse:o,stringify:r}},function(e,t){function a(e,t,n,r,o,i,s){try{var u=e[i](s),a=u.value}catch(e){return void n(e)}u.done?t(a):Promise.resolve(a).then(r,o)}e.exports=function(u){return function(){var e=this,s=arguments;return new Promise(function(t,n){var r=u.apply(e,s);function o(e){a(r,t,n,o,i,"next",e)}function i(e){a(r,t,n,o,i,"throw",e)}o(void 0)})}}},function(e,t,n){"use strict";var s=Object.prototype.hasOwnProperty,u=function(){for(var e=[],t=0;t<256;++t)e.push("%"+((t<16?"0":"")+t.toString(16)).toUpperCase());return e}(),a=function(e,t){for(var n=t&&t.plainObjects?Object.create(null):{},r=0;r<e.length;++r)void 0!==e[r]&&(n[r]=e[r]);return n};e.exports={arrayToObject:a,assign:function(e,n){return Object.keys(n).reduce(function(e,t){return e[t]=n[t],e},e)},combine:function(e,t){return[].concat(e,t)},compact:function(e){for(var t=[{obj:{o:e},prop:"o"}],n=[],r=0;r<t.length;++r)for(var o=t[r],i=o.obj[o.prop],s=Object.keys(i),u=0;u<s.length;++u){var a=s[u],c=i[a];"object"==typeof c&&null!==c&&-1===n.indexOf(c)&&(t.push({obj:i,prop:a}),n.push(c))}return function(e){for(;1<e.length;){var t=e.pop(),n=t.obj[t.prop];if(Array.isArray(n)){for(var r=[],o=0;o<n.length;++o)void 0!==n[o]&&r.push(n[o]);t.obj[t.prop]=r}}}(t),e},decode:function(e,t,n){var r=e.replace(/\+/g," ");if("iso-8859-1"===n)return r.replace(/%[0-9a-f]{2}/gi,unescape);try{return decodeURIComponent(r)}catch(e){return r}},encode:function(e,t,n){if(0===e.length)return e;var r="string"==typeof e?e:String(e);if("iso-8859-1"===n)return escape(r).replace(/%u[0-9a-f]{4}/gi,function(e){return"%26%23"+parseInt(e.slice(2),16)+"%3B"});for(var o="",i=0;i<r.length;++i){var s=r.charCodeAt(i);45===s||46===s||95===s||126===s||48<=s&&s<=57||65<=s&&s<=90||97<=s&&s<=122?o+=r.charAt(i):s<128?o+=u[s]:s<2048?o+=u[192|s>>6]+u[128|63&s]:s<55296||57344<=s?o+=u[224|s>>12]+u[128|s>>6&63]+u[128|63&s]:(i+=1,s=65536+((1023&s)<<10|1023&r.charCodeAt(i)),o+=u[240|s>>18]+u[128|s>>12&63]+u[128|s>>6&63]+u[128|63&s])}return o},isBuffer:function(e){return null!=e&&!!(e.constructor&&e.constructor.isBuffer&&e.constructor.isBuffer(e))},isRegExp:function(e){return"[object RegExp]"===Object.prototype.toString.call(e)},merge:function r(n,o,i){if(!o)return n;if("object"!=typeof o){if(Array.isArray(n))n.push(o);else{if("object"!=typeof n)return[n,o];(i&&(i.plainObjects||i.allowPrototypes)||!s.call(Object.prototype,o))&&(n[o]=!0)}return n}if("object"!=typeof n)return[n].concat(o);var e=n;return Array.isArray(n)&&!Array.isArray(o)&&(e=a(n,i)),Array.isArray(n)&&Array.isArray(o)?(o.forEach(function(e,t){s.call(n,t)?n[t]&&"object"==typeof n[t]?n[t]=r(n[t],e,i):n.push(e):n[t]=e}),n):Object.keys(o).reduce(function(e,t){var n=o[t];return s.call(e,t)?e[t]=r(e[t],n,i):e[t]=n,e},e)}}},function(e,t,n){"use strict";var r=String.prototype.replace,o=/%20/g;e.exports={default:"RFC3986",formatters:{RFC1738:function(e){return r.call(e,o,"+")},RFC3986:function(e){return e}},RFC1738:"RFC1738",RFC3986:"RFC3986"}},function(e,t){!function(){e.exports=this.wp.blocks}()},function(e,t){!function(){e.exports=this.wp.editor}()},function(e,t,n){e.exports=n(26)},function(e,t){!function(){e.exports=this.lodash}()},function(t,e){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function r(e){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?t.exports=r=function(e){return n(e)}:t.exports=r=function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":n(e)},r(e)}t.exports=r},function(n,e){function r(e,t){return n.exports=r=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e},r(e,t)}n.exports=r},function(e,t,n){"use strict";var x=n(15),S=n(16),E={brackets:function(e){return e+"[]"},indices:function(e,t){return e+"["+t+"]"},repeat:function(e){return e}},r=Array.isArray,o=Array.prototype.push,P=function(e,t){o.apply(e,r(t)?t:[t])},i=Date.prototype.toISOString,C={addQueryPrefix:!1,allowDots:!1,charset:"utf-8",charsetSentinel:!1,delimiter:"&",encode:!0,encoder:x.encode,encodeValuesOnly:!1,indices:!1,serializeDate:function(e){return i.call(e)},skipNulls:!1,strictNullHandling:!1},k=function e(t,n,r,o,i,s,u,a,c,l,p,f,d){var h=t;if("function"==typeof u?h=u(n,h):h instanceof Date&&(h=l(h)),null===h){if(o)return s&&!f?s(n,C.encoder,d):n;h=""}if("string"==typeof h||"number"==typeof h||"boolean"==typeof h||x.isBuffer(h))return s?[p(f?n:s(n,C.encoder,d))+"="+p(s(h,C.encoder,d))]:[p(n)+"="+p(String(h))];var m,y=[];if(void 0===h)return y;if(Array.isArray(u))m=u;else{var b=Object.keys(h);m=a?b.sort(a):b}for(var v=0;v<m.length;++v){var g=m[v];i&&null===h[g]||(Array.isArray(h)?P(y,e(h[g],r(n,g),r,o,i,s,u,a,c,l,p,f,d)):P(y,e(h[g],n+(c?"."+g:"["+g+"]"),r,o,i,s,u,a,c,l,p,f,d)))}return y};e.exports=function(e,t){var n=e,r=t?x.assign({},t):{};if(null!==r.encoder&&void 0!==r.encoder&&"function"!=typeof r.encoder)throw new TypeError("Encoder has to be a function.");var o=void 0===r.delimiter?C.delimiter:r.delimiter,i="boolean"==typeof r.strictNullHandling?r.strictNullHandling:C.strictNullHandling,s="boolean"==typeof r.skipNulls?r.skipNulls:C.skipNulls,u="boolean"==typeof r.encode?r.encode:C.encode,a="function"==typeof r.encoder?r.encoder:C.encoder,c="function"==typeof r.sort?r.sort:null,l=void 0===r.allowDots?C.allowDots:!!r.allowDots,p="function"==typeof r.serializeDate?r.serializeDate:C.serializeDate,f="boolean"==typeof r.encodeValuesOnly?r.encodeValuesOnly:C.encodeValuesOnly,d=r.charset||C.charset;if(void 0!==r.charset&&"utf-8"!==r.charset&&"iso-8859-1"!==r.charset)throw new Error("The charset option must be either utf-8, iso-8859-1, or undefined");if(void 0===r.format)r.format=S.default;else if(!Object.prototype.hasOwnProperty.call(S.formatters,r.format))throw new TypeError("Unknown format option provided.");var h,m,y=S.formatters[r.format];"function"==typeof r.filter?n=(m=r.filter)("",n):Array.isArray(r.filter)&&(h=m=r.filter);var b,v=[];if("object"!=typeof n||null===n)return"";b=r.arrayFormat in E?r.arrayFormat:"indices"in r?r.indices?"indices":"repeat":"indices";var g=E[b];h||(h=Object.keys(n)),c&&h.sort(c);for(var O=0;O<h.length;++O){var w=h[O];s&&null===n[w]||P(v,k(n[w],w,g,i,s,u?a:null,m,c,l,p,y,f,d))}var j=v.join(o),_=!0===r.addQueryPrefix?"?":"";return r.charsetSentinel&&(_+="iso-8859-1"===d?"utf8=%26%2310003%3B&":"utf8=%E2%9C%93&"),0<j.length?_+j:""}},function(e,t,n){"use strict";var h=n(15),m=Object.prototype.hasOwnProperty,y={allowDots:!1,allowPrototypes:!1,arrayLimit:20,charset:"utf-8",charsetSentinel:!1,decoder:h.decode,delimiter:"&",depth:5,ignoreQueryPrefix:!1,interpretNumericEntities:!1,parameterLimit:1e3,parseArrays:!0,plainObjects:!1,strictNullHandling:!1},c=function(e,t,n){if(e){var r=n.allowDots?e.replace(/\.([^.[]+)/g,"[$1]"):e,o=/(\[[^[\]]*])/g,i=/(\[[^[\]]*])/.exec(r),s=i?r.slice(0,i.index):r,u=[];if(s){if(!n.plainObjects&&m.call(Object.prototype,s)&&!n.allowPrototypes)return;u.push(s)}for(var a=0;null!==(i=o.exec(r))&&a<n.depth;){if(a+=1,!n.plainObjects&&m.call(Object.prototype,i[1].slice(1,-1))&&!n.allowPrototypes)return;u.push(i[1])}return i&&u.push("["+r.slice(i.index)+"]"),function(e,t,n){for(var r=t,o=e.length-1;0<=o;--o){var i,s=e[o];if("[]"===s&&n.parseArrays)i=[].concat(r);else{i=n.plainObjects?Object.create(null):{};var u="["===s.charAt(0)&&"]"===s.charAt(s.length-1)?s.slice(1,-1):s,a=parseInt(u,10);n.parseArrays||""!==u?!isNaN(a)&&s!==u&&String(a)===u&&0<=a&&n.parseArrays&&a<=n.arrayLimit?(i=[])[a]=r:i[u]=r:i={0:r}}r=i}return r}(u,t,n)}};e.exports=function(e,t){var n=t?h.assign({},t):{};if(null!==n.decoder&&void 0!==n.decoder&&"function"!=typeof n.decoder)throw new TypeError("Decoder has to be a function.");if(n.ignoreQueryPrefix=!0===n.ignoreQueryPrefix,n.delimiter="string"==typeof n.delimiter||h.isRegExp(n.delimiter)?n.delimiter:y.delimiter,n.depth="number"==typeof n.depth?n.depth:y.depth,n.arrayLimit="number"==typeof n.arrayLimit?n.arrayLimit:y.arrayLimit,n.parseArrays=!1!==n.parseArrays,n.decoder="function"==typeof n.decoder?n.decoder:y.decoder,n.allowDots=void 0===n.allowDots?y.allowDots:!!n.allowDots,n.plainObjects="boolean"==typeof n.plainObjects?n.plainObjects:y.plainObjects,n.allowPrototypes="boolean"==typeof n.allowPrototypes?n.allowPrototypes:y.allowPrototypes,n.parameterLimit="number"==typeof n.parameterLimit?n.parameterLimit:y.parameterLimit,n.strictNullHandling="boolean"==typeof n.strictNullHandling?n.strictNullHandling:y.strictNullHandling,void 0!==n.charset&&"utf-8"!==n.charset&&"iso-8859-1"!==n.charset)throw new Error("The charset option must be either utf-8, iso-8859-1, or undefined");if(void 0===n.charset&&(n.charset=y.charset),""===e||null==e)return n.plainObjects?Object.create(null):{};for(var r="string"==typeof e?function(e,t){var n,r={},o=t.ignoreQueryPrefix?e.replace(/^\?/,""):e,i=t.parameterLimit===1/0?void 0:t.parameterLimit,s=o.split(t.delimiter,i),u=-1,a=t.charset;if(t.charsetSentinel)for(n=0;n<s.length;++n)0===s[n].indexOf("utf8=")&&("utf8=%E2%9C%93"===s[n]?a="utf-8":"utf8=%26%2310003%3B"===s[n]&&(a="iso-8859-1"),u=n,n=s.length);for(n=0;n<s.length;++n)if(n!==u){var c,l,p=s[n],f=p.indexOf("]="),d=-1===f?p.indexOf("="):f+1;(l=-1===d?(c=t.decoder(p,y.decoder,a),t.strictNullHandling?null:""):(c=t.decoder(p.slice(0,d),y.decoder,a),t.decoder(p.slice(d+1),y.decoder,a)))&&t.interpretNumericEntities&&"iso-8859-1"===a&&(l=l.replace(/&#(\d+);/g,function(e,t){return String.fromCharCode(parseInt(t,10))})),m.call(r,c)?r[c]=h.combine(r[c],l):r[c]=l}return r}(e,n):e,o=n.plainObjects?Object.create(null):{},i=Object.keys(r),s=0;s<i.length;++s){var u=i[s],a=c(u,r[u],n);o=h.merge(o,a,n)}return h.compact(o)}},function(e,t,n){},function(e,t,n){var r;window,r=function(n){return function(n){var r={};function o(e){if(r[e])return r[e].exports;var t=r[e]={i:e,l:!1,exports:{}};return n[e].call(t.exports,t,t.exports,o),t.l=!0,t.exports}return o.m=n,o.c=r,o.d=function(e,t,n){o.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.t=function(t,e){if(1&e&&(t=o(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(o.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)o.d(n,r,function(e){return t[e]}.bind(null,r));return n},o.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(t,"a",t),t},o.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},o.p="/",o(o.s=31)}([function(e,t){var n=e.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=n)},function(e,t,n){var m=n(0),y=n(5),b=n(6),v=n(14),g=n(16),O="prototype",w=function(e,t,n){var r,o,i,s,u=e&w.F,a=e&w.G,c=e&w.S,l=e&w.P,p=e&w.B,f=a?m:c?m[t]||(m[t]={}):(m[t]||{})[O],d=a?y:y[t]||(y[t]={}),h=d[O]||(d[O]={});for(r in a&&(n=t),n)i=((o=!u&&f&&void 0!==f[r])?f:n)[r],s=p&&o?g(i,m):l&&"function"==typeof i?g(Function.call,i):i,f&&v(f,r,i,e&w.U),d[r]!=i&&b(d,r,s),l&&h[r]!=i&&(h[r]=i)};m.core=y,w.F=1,w.G=2,w.S=4,w.P=8,w.B=16,w.W=32,w.U=64,w.R=128,e.exports=w},function(e,t){e.exports=function(e){return"object"==typeof e?null!==e:"function"==typeof e}},function(e,t,n){e.exports=!n(4)(function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})},function(e,t){e.exports=function(e){try{return!!e()}catch(e){return!0}}},function(e,t){var n=e.exports={version:"2.5.7"};"number"==typeof __e&&(__e=n)},function(e,t,n){var r=n(7),o=n(35);e.exports=n(3)?function(e,t,n){return r.f(e,t,o(1,n))}:function(e,t,n){return e[t]=n,e}},function(e,t,n){var r=n(8),o=n(33),i=n(34),s=Object.defineProperty;t.f=n(3)?Object.defineProperty:function(e,t,n){if(r(e),t=i(t,!0),r(n),o)try{return s(e,t,n)}catch(e){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(e[t]=n.value),e}},function(e,t,n){var r=n(2);e.exports=function(e){if(!r(e))throw TypeError(e+" is not an object!");return e}},function(e,t){var n=0,r=Math.random();e.exports=function(e){return"Symbol(".concat(void 0===e?"":e,")_",(++n+r).toString(36))}},function(e,t,n){var r=n(20);e.exports=Object("z").propertyIsEnumerable(0)?Object:function(e){return"String"==r(e)?e.split(""):Object(e)}},function(e,t){e.exports=function(e){if(null==e)throw TypeError("Can't call method on  "+e);return e}},function(e,t){e.exports=n},function(e,t,n){var r=n(2),o=n(0).document,i=r(o)&&r(o.createElement);e.exports=function(e){return i?o.createElement(e):{}}},function(e,t,n){var i=n(0),s=n(6),u=n(15),a=n(9)("src"),r="toString",o=Function[r],c=(""+o).split(r);n(5).inspectSource=function(e){return o.call(e)},(e.exports=function(e,t,n,r){var o="function"==typeof n;o&&(u(n,"name")||s(n,"name",t)),e[t]!==n&&(o&&(u(n,a)||s(n,a,e[t]?""+e[t]:c.join(String(t)))),e===i?e[t]=n:r?e[t]?e[t]=n:s(e,t,n):(delete e[t],s(e,t,n)))})(Function.prototype,r,function(){return"function"==typeof this&&this[a]||o.call(this)})},function(e,t){var n={}.hasOwnProperty;e.exports=function(e,t){return n.call(e,t)}},function(e,t,n){var i=n(17);e.exports=function(r,o,e){if(i(r),void 0===o)return r;switch(e){case 1:return function(e){return r.call(o,e)};case 2:return function(e,t){return r.call(o,e,t)};case 3:return function(e,t,n){return r.call(o,e,t,n)}}return function(){return r.apply(o,arguments)}}},function(e,t){e.exports=function(e){if("function"!=typeof e)throw TypeError(e+" is not a function!");return e}},function(e,t,n){var r=n(37),o=n(26);e.exports=Object.keys||function(e){return r(e,o)}},function(e,t,n){var r=n(10),o=n(11);e.exports=function(e){return r(o(e))}},function(e,t){var n={}.toString;e.exports=function(e){return n.call(e).slice(8,-1)}},function(e,t,n){var a=n(19),c=n(22),l=n(38);e.exports=function(u){return function(e,t,n){var r,o=a(e),i=c(o.length),s=l(n,i);if(u&&t!=t){for(;s<i;)if((r=o[s++])!=r)return!0}else for(;s<i;s++)if((u||s in o)&&o[s]===t)return u||s||0;return!u&&-1}}},function(e,t,n){var r=n(23),o=Math.min;e.exports=function(e){return 0<e?o(r(e),9007199254740991):0}},function(e,t){var n=Math.ceil,r=Math.floor;e.exports=function(e){return isNaN(e=+e)?0:(0<e?r:n)(e)}},function(e,t,n){var r=n(25)("keys"),o=n(9);e.exports=function(e){return r[e]||(r[e]=o(e))}},function(e,t,n){var r=n(5),o=n(0),i="__core-js_shared__",s=o[i]||(o[i]={});(e.exports=function(e,t){return s[e]||(s[e]=void 0!==t?t:{})})("versions",[]).push({version:r.version,mode:n(39)?"pure":"global",copyright:"© 2018 Denis Pushkarev (zloirock.ru)"})},function(e,t){e.exports="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(",")},function(e,t,n){var r=n(11);e.exports=function(e){return Object(r(e))}},function(e,t,n){var r=n(1);r(r.S,"Object",{create:n(42)})},function(e,t,n){var r=n(25)("wks"),o=n(9),i=n(0).Symbol,s="function"==typeof i;(e.exports=function(e){return r[e]||(r[e]=s&&i[e]||(s?i:o)("Symbol."+e))}).store=r},function(e,t,n){"use strict";var r=n(4);e.exports=function(e,t){return!!e&&r(function(){t?e.call(null,function(){},1):e.call(null)})}},function(e,t,n){"use strict";t.__esModule=!0,t.default=void 0,n(32),n(28),n(45),n(50),n(51),n(52),n(55);var H=n(12),$=o(n(57)),r=o(n(58));function o(e){return e&&e.__esModule?e:{default:e}}function K(){return(K=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e}).apply(this,arguments)}function i(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}var s,u={13:"enter",27:"escape",32:"space",38:"up",40:"down"},z=((s=document.createElement("x")).style.cssText="pointer-events:auto","auto"===s.style.pointerEvents);function f(){return!(!navigator.userAgent.match(/(iPod|iPhone|iPad)/g)||!navigator.userAgent.match(/AppleWebKit/g))}var a=function(n){function e(e){var t;return(t=n.call(this,e)||this).elementReferences={},t.state={focused:null,hovered:null,clicked:null,menuOpen:!1,options:e.defaultValue?[e.defaultValue]:[],query:e.defaultValue,selected:null},t.handleComponentBlur=t.handleComponentBlur.bind(i(i(t))),t.handleKeyDown=t.handleKeyDown.bind(i(i(t))),t.handleUpArrow=t.handleUpArrow.bind(i(i(t))),t.handleDownArrow=t.handleDownArrow.bind(i(i(t))),t.handleEnter=t.handleEnter.bind(i(i(t))),t.handlePrintableKey=t.handlePrintableKey.bind(i(i(t))),t.handleListMouseLeave=t.handleListMouseLeave.bind(i(i(t))),t.handleOptionBlur=t.handleOptionBlur.bind(i(i(t))),t.handleOptionClick=t.handleOptionClick.bind(i(i(t))),t.handleOptionFocus=t.handleOptionFocus.bind(i(i(t))),t.handleOptionMouseEnter=t.handleOptionMouseEnter.bind(i(i(t))),t.handleInputBlur=t.handleInputBlur.bind(i(i(t))),t.handleInputChange=t.handleInputChange.bind(i(i(t))),t.handleInputFocus=t.handleInputFocus.bind(i(i(t))),t.pollInputElement=t.pollInputElement.bind(i(i(t))),t.getDirectInputChanges=t.getDirectInputChanges.bind(i(i(t))),t}var t,r;r=n,(t=e).prototype=Object.create(r.prototype),(t.prototype.constructor=t).__proto__=r;var o=e.prototype;return o.componentDidMount=function(){this.pollInputElement()},o.componentWillUnmount=function(){clearTimeout(this.$pollInput),clearTimeout(this.$blurInput)},o.pollInputElement=function(){var e=this;this.getDirectInputChanges(),this.$pollInput=setTimeout(function(){e.pollInputElement()},100)},o.getDirectInputChanges=function(){var e=this.elementReferences[-1];e&&e.value!==this.state.query&&this.handleInputChange({target:{value:e.value}})},o.componentDidUpdate=function(e,t){var n=this.state,r=n.focused,o=n.clicked,i=null===r,s=t.focused!==r;(s&&!i||null!==o)&&this.elementReferences[r].focus();var u=-1===r,a=s&&null===t.focused;if(u&&a){var c=this.elementReferences[r];c.setSelectionRange(0,c.value.length)}},o.hasAutoselect=function(){return!f()&&this.props.autoselect},o.templateInputValue=function(e){var t=this.props.templates&&this.props.templates.inputValue;return t?t(e):e},o.templateSuggestion=function(e){var t=this.props.templates&&this.props.templates.suggestion;return t?t(e):e},o.handleComponentBlur=function(e){var t,n=this.state,r=n.options,o=n.query,i=n.selected;this.props.confirmOnBlur?(t=e.query||o,this.props.onConfirm(r[i])):t=o,this.setState({focused:null,clicked:null,menuOpen:e.menuOpen||!1,query:t,selected:null})},o.handleListMouseLeave=function(e){this.setState({hovered:null})},o.handleOptionBlur=function(e,t){var n=this.state,r=n.focused,o=n.clicked,i=n.menuOpen,s=n.options,u=n.selected,a=null===e.relatedTarget&&null===o,c=e.relatedTarget===this.elementReferences[-1],l=r!==t&&-1!==r;if(!l&&a||!l&&!c){var p=i&&f();this.handleComponentBlur({menuOpen:p,query:this.templateInputValue(s[u])})}},o.handleInputBlur=function(e){var t=this,n=this.state,r=n.focused,o=n.menuOpen,i=n.options,s=n.query,u=n.selected,a=-1!==r;if(clearTimeout(this.$blurInput),!a){var c=o&&f(),l=f()?s:this.templateInputValue(i[u]);this.$blurInput=setTimeout(function(){return t.handleComponentBlur({menuOpen:c,query:l})},200)}},o.handleInputChange=function(e){var n=this,t=this.props,r=t.minLength,o=t.source,i=t.showAllValues,s=this.hasAutoselect(),u=e.target.value,a=0===u.length,c=this.state.query.length!==u.length,l=u.length>=r;this.setState({query:u}),i||!a&&c&&l?o(u,function(e){var t=0<e.length;n.setState({menuOpen:t,options:e,selected:s&&t?0:-1})}):!a&&l||this.setState({menuOpen:!1,options:[]})},o.handleInputClick=function(e){this.handleInputChange(e)},o.handleInputFocus=function(e){this.setState({focused:-1})},o.handleOptionFocus=function(e){this.setState({focused:e,hovered:null,selected:e})},o.handleOptionMouseEnter=function(e,t){f()||this.setState({hovered:t})},o.handleOptionClick=function(e,t){var n=this.state.options[t],r=this.templateInputValue(n);clearTimeout(this.$blurInput),this.props.onConfirm(n),this.setState({focused:-1,clicked:t,hovered:null,menuOpen:!1,query:r,selected:-1}),this.forceUpdate()},o.handleUpArrow=function(e){e.preventDefault();var t=this.state,n=t.menuOpen,r=t.selected;-1!==r&&n&&this.handleOptionFocus(r-1)},o.handleDownArrow=function(e){var t=this;if(e.preventDefault(),this.props.showAllValues&&!1===this.state.menuOpen)e.preventDefault(),this.props.source("",function(e){t.setState({menuOpen:!0,options:e,selected:0,focused:0,hovered:null})});else if(!0===this.state.menuOpen){var n=this.state,r=n.menuOpen,o=n.options,i=n.selected;i!==o.length-1&&r&&this.handleOptionFocus(i+1)}},o.handleSpace=function(e){var t=this;this.props.showAllValues&&!1===this.state.menuOpen&&""===this.state.query&&(e.preventDefault(),this.props.source("",function(e){t.setState({menuOpen:!0,options:e})})),-1!==this.state.focused&&(e.preventDefault(),this.handleOptionClick(e,this.state.focused))},o.handleEnter=function(e){this.state.menuOpen&&(e.preventDefault(),0<=this.state.selected&&this.handleOptionClick(e,this.state.selected))},o.handlePrintableKey=function(e){var t=this.elementReferences[-1];e.target===t||t.focus()},o.handleKeyDown=function(e){switch(u[e.keyCode]){case"up":this.handleUpArrow(e);break;case"down":this.handleDownArrow(e);break;case"space":this.handleSpace(e);break;case"enter":this.handleEnter(e);break;case"escape":this.handleComponentBlur({query:this.state.query});break;default:(47<(t=e.keyCode)&&t<58||32===t||8===t||64<t&&t<91||95<t&&t<112||185<t&&t<193||218<t&&t<223)&&this.handlePrintableKey(e)}var t},o.render=function(){var e,o=this,t=this.props,n=t.cssNamespace,r=t.displayMenu,i=t.id,s=t.minLength,u=t.name,a=t.placeholder,c=t.required,l=t.showAllValues,p=t.tNoResults,f=t.tStatusQueryTooShort,d=t.tStatusNoResults,h=t.tStatusSelectedOption,m=t.tStatusResults,y=t.dropdownArrow,b=this.state,v=b.focused,g=b.hovered,O=b.menuOpen,w=b.options,j=b.query,_=b.selected,x=this.hasAutoselect(),S=-1===v,E=0===w.length,P=0!==j.length,C=j.length>=s,k=this.props.showNoOptionsFound&&S&&E&&P&&C,A=n+"__wrapper",I=n+"__input",N=null!==v?" "+I+"--focused":"",L=this.props.showAllValues?" "+I+"--show-all-values":" "+I+"--default",T=n+"__dropdown-arrow-down",D=-1!==v&&null!==v,R=n+"__menu",F=R+"--"+r,B=R+"--"+(O||k?"visible":"hidden"),M=n+"__option",q=n+"__hint",V=this.templateInputValue(w[_]),U=V&&0===V.toLowerCase().indexOf(j.toLowerCase())&&x?j+V.substr(j.length):"",Q=z&&U;return l&&"string"==typeof(e=y({className:T}))&&(e=(0,H.createElement)("div",{className:n+"__dropdown-arrow-down-wrapper",dangerouslySetInnerHTML:{__html:e}})),(0,H.createElement)("div",{className:A,onKeyDown:this.handleKeyDown,role:"combobox","aria-expanded":O?"true":"false"},(0,H.createElement)($.default,{length:w.length,queryLength:j.length,minQueryLength:s,selectedOption:this.templateInputValue(w[_]),selectedOptionIndex:_,tQueryTooShort:f,tNoResults:d,tSelectedOption:h,tResults:m}),Q&&(0,H.createElement)("span",null,(0,H.createElement)("input",{className:q,readonly:!0,tabIndex:"-1",value:U})),(0,H.createElement)("input",K({"aria-activedescendant":!!D&&i+"__option--"+v,"aria-owns":i+"__listbox",autoComplete:"off",className:""+I+N+L,id:i,onClick:function(e){return o.handleInputClick(e)},onBlur:this.handleInputBlur},{onChange:this.handleInputChange},{onFocus:this.handleInputFocus,name:u,placeholder:a,ref:function(e){o.elementReferences[-1]=e},type:"text",role:"textbox",required:c,value:j})),e,(0,H.createElement)("ul",{className:R+" "+F+" "+B,onMouseLeave:function(e){return o.handleListMouseLeave(e)},id:i+"__listbox",role:"listbox"},w.map(function(e,t){var n=(-1===v?_===t:v===t)&&null===g?" "+M+"--focused":"",r=t%2?" "+M+"--odd":"";return(0,H.createElement)("li",{"aria-selected":v===t,className:""+M+n+r,dangerouslySetInnerHTML:{__html:o.templateSuggestion(e)},id:i+"__option--"+t,key:t,onBlur:function(e){return o.handleOptionBlur(e,t)},onClick:function(e){return o.handleOptionClick(e,t)},onMouseEnter:function(e){return o.handleOptionMouseEnter(e,t)},ref:function(e){o.elementReferences[t]=e},role:"option",tabIndex:"-1"})}),k&&(0,H.createElement)("li",{className:M+" "+M+"--no-results"},p())))},e}(H.Component);(t.default=a).defaultProps={autoselect:!1,cssNamespace:"autocomplete",defaultValue:"",displayMenu:"inline",minLength:0,name:"input-autocomplete",placeholder:"",onConfirm:function(){},confirmOnBlur:!0,showNoOptionsFound:!0,showAllValues:!1,required:!1,tNoResults:function(){return"No results found"},dropdownArrow:r.default}},function(e,t,n){var r=n(1);r(r.S+r.F,"Object",{assign:n(36)})},function(e,t,n){e.exports=!n(3)&&!n(4)(function(){return 7!=Object.defineProperty(n(13)("div"),"a",{get:function(){return 7}}).a})},function(e,t,n){var o=n(2);e.exports=function(e,t){if(!o(e))return e;var n,r;if(t&&"function"==typeof(n=e.toString)&&!o(r=n.call(e)))return r;if("function"==typeof(n=e.valueOf)&&!o(r=n.call(e)))return r;if(!t&&"function"==typeof(n=e.toString)&&!o(r=n.call(e)))return r;throw TypeError("Can't convert object to primitive value")}},function(e,t){e.exports=function(e,t){return{enumerable:!(1&e),configurable:!(2&e),writable:!(4&e),value:t}}},function(e,t,n){"use strict";var f=n(18),d=n(40),h=n(41),m=n(27),y=n(10),o=Object.assign;e.exports=!o||n(4)(function(){var e={},t={},n=Symbol(),r="abcdefghijklmnopqrst";return e[n]=7,r.split("").forEach(function(e){t[e]=e}),7!=o({},e)[n]||Object.keys(o({},t)).join("")!=r})?function(e,t){for(var n=m(e),r=arguments.length,o=1,i=d.f,s=h.f;o<r;)for(var u,a=y(arguments[o++]),c=i?f(a).concat(i(a)):f(a),l=c.length,p=0;p<l;)s.call(a,u=c[p++])&&(n[u]=a[u]);return n}:o},function(e,t,n){var s=n(15),u=n(19),a=n(21)(!1),c=n(24)("IE_PROTO");e.exports=function(e,t){var n,r=u(e),o=0,i=[];for(n in r)n!=c&&s(r,n)&&i.push(n);for(;t.length>o;)s(r,n=t[o++])&&(~a(i,n)||i.push(n));return i}},function(e,t,n){var r=n(23),o=Math.max,i=Math.min;e.exports=function(e,t){return(e=r(e))<0?o(e+t,0):i(e,t)}},function(e,t){e.exports=!1},function(e,t){t.f=Object.getOwnPropertySymbols},function(e,t){t.f={}.propertyIsEnumerable},function(e,t,r){var o=r(8),i=r(43),s=r(26),u=r(24)("IE_PROTO"),a=function(){},c="prototype",l=function(){var e,t=r(13)("iframe"),n=s.length;for(t.style.display="none",r(44).appendChild(t),t.src="javascript:",(e=t.contentWindow.document).open(),e.write("<script>document.F=Object<\/script>"),e.close(),l=e.F;n--;)delete l[c][s[n]];return l()};e.exports=Object.create||function(e,t){var n;return null!==e?(a[c]=o(e),n=new a,a[c]=null,n[u]=e):n=l(),void 0===t?n:i(n,t)}},function(e,t,n){var s=n(7),u=n(8),a=n(18);e.exports=n(3)?Object.defineProperties:function(e,t){u(e);for(var n,r=a(t),o=r.length,i=0;i<o;)s.f(e,n=r[i++],t[n]);return e}},function(e,t,n){var r=n(0).document;e.exports=r&&r.documentElement},function(e,t,n){"use strict";var r=n(1),o=n(46)(1);r(r.P+r.F*!n(30)([].map,!0),"Array",{map:function(e){return o(this,e,arguments[1])}})},function(e,t,n){var g=n(16),O=n(10),w=n(27),j=n(22),r=n(47);e.exports=function(p,e){var f=1==p,d=2==p,h=3==p,m=4==p,y=6==p,b=5==p||y,v=e||r;return function(e,t,n){for(var r,o,i=w(e),s=O(i),u=g(t,n,3),a=j(s.length),c=0,l=f?v(e,a):d?v(e,0):void 0;c<a;c++)if((b||c in s)&&(o=u(r=s[c],c,i),p))if(f)l[c]=o;else if(o)switch(p){case 3:return!0;case 5:return r;case 6:return c;case 2:l.push(r)}else if(m)return!1;return y?-1:h||m?m:l}}},function(e,t,n){var r=n(48);e.exports=function(e,t){return new(r(e))(t)}},function(e,t,n){var r=n(2),o=n(49),i=n(29)("species");e.exports=function(e){var t;return o(e)&&("function"!=typeof(t=e.constructor)||t!==Array&&!o(t.prototype)||(t=void 0),r(t)&&null===(t=t[i])&&(t=void 0)),void 0===t?Array:t}},function(e,t,n){var r=n(20);e.exports=Array.isArray||function(e){return"Array"==r(e)}},function(e,t,n){"use strict";var r=n(1),o=n(21)(!1),i=[].indexOf,s=!!i&&1/[1].indexOf(1,-0)<0;r(r.P+r.F*(s||!n(30)(i)),"Array",{indexOf:function(e){return s?i.apply(this,arguments)||0:o(this,e,arguments[1])}})},function(e,t,n){var r=n(7).f,o=Function.prototype,i=/^\s*function ([^ (]*)/;"name"in o||n(3)&&r(o,"name",{configurable:!0,get:function(){try{return(""+this).match(i)[1]}catch(e){return""}}})},function(e,t,n){var r=n(1);r(r.P,"Function",{bind:n(53)})},function(e,t,n){"use strict";var i=n(17),s=n(2),u=n(54),a=[].slice,c={};e.exports=Function.bind||function(t){var n=i(this),r=a.call(arguments,1),o=function(){var e=r.concat(a.call(arguments));return this instanceof o?function(e,t,n){if(!(t in c)){for(var r=[],o=0;o<t;o++)r[o]="a["+o+"]";c[t]=Function("F,a","return new F("+r.join(",")+")")}return c[t](e,n)}(n,e.length,e):u(n,e,t)};return s(n.prototype)&&(o.prototype=n.prototype),o}},function(e,t){e.exports=function(e,t,n){var r=void 0===n;switch(t.length){case 0:return r?e():e.call(n);case 1:return r?e(t[0]):e.call(n,t[0]);case 2:return r?e(t[0],t[1]):e.call(n,t[0],t[1]);case 3:return r?e(t[0],t[1],t[2]):e.call(n,t[0],t[1],t[2]);case 4:return r?e(t[0],t[1],t[2],t[3]):e.call(n,t[0],t[1],t[2],t[3])}return e.apply(n,t)}},function(e,t,n){n(56)("match",1,function(r,o,e){return[function(e){"use strict";var t=r(this),n=null==e?void 0:e[o];return void 0!==n?n.call(e,t):new RegExp(e)[o](String(t))},e]})},function(e,t,n){"use strict";var u=n(6),a=n(14),c=n(4),l=n(11),p=n(29);e.exports=function(t,e,n){var r=p(t),o=n(l,r,""[t]),i=o[0],s=o[1];c(function(){var e={};return e[r]=function(){return 7},7!=""[t](e)})&&(a(String.prototype,t,i),u(RegExp.prototype,r,2==e?function(e,t){return s.call(e,this,t)}:function(e){return s.call(e,this)}))}},function(e,t,n){"use strict";t.__esModule=!0,t.default=void 0,n(28);var m=n(12),r=function(o){function e(){for(var e,t=arguments.length,n=new Array(t),r=0;r<t;r++)n[r]=arguments[r];return(e=o.call.apply(o,[this].concat(n))||this).state={bump:!1},e}var t,n;n=o,(t=e).prototype=Object.create(n.prototype),(t.prototype.constructor=t).__proto__=n;var r=e.prototype;return r.componentWillReceiveProps=function(e){e.queryLength!==this.props.queryLength&&this.setState(function(e){return{bump:!e.bump}})},r.render=function(){var e,t=this.props,n=t.length,r=t.queryLength,o=t.minQueryLength,i=t.selectedOption,s=t.selectedOptionIndex,u=t.tQueryTooShort,a=t.tNoResults,c=t.tSelectedOption,l=t.tResults,p=this.state.bump,f=r<o,d=0===n,h=i?c(i,n,s):"";return e=f?u(o):d?a():l(n,h),(0,m.createElement)("div",{"aria-atomic":"true","aria-live":"polite",role:"status",style:{border:"0",clip:"rect(0 0 0 0)",height:"1px",marginBottom:"-1px",marginRight:"-1px",overflow:"hidden",padding:"0",position:"absolute",whiteSpace:"nowrap",width:"1px"}},e,(0,m.createElement)("span",null,p?",":",,"))},e}(m.Component);(t.default=r).defaultProps={tQueryTooShort:function(e){return"Type in "+e+" or more characters for results."},tNoResults:function(){return"No search results."},tSelectedOption:function(e,t,n){return e+" ("+(n+1)+" of "+t+") is selected."},tResults:function(e,t){return e+" "+(1===e?"result":"results")+" "+(1===e?"is":"are")+" available. "+t}}},function(e,t,n){"use strict";t.__esModule=!0,t.default=void 0;var r=n(12);t.default=function(e){var t=e.className;return(0,r.createElement)("svg",{version:"1.1",xmlns:"http://www.w3.org/2000/svg",className:t,focusable:"false"},(0,r.createElement)("g",{stroke:"none",fill:"none","fill-rule":"evenodd"},(0,r.createElement)("polygon",{fill:"#000000",points:"0 0 22 0 11 17"})))}}])},e.exports=r(n(27))},function(e,t){!function(){e.exports=this.React}()},function(e,t,n){},function(e,t,n){"use strict";n.r(t);var r=n(17),o=n(2),f=n(1),i=n(4),s=n.n(i),u=n(5),a=n.n(u),c=n(6),l=n.n(c),p=n(7),d=n.n(p),h=n(8),m=n.n(h),y=n(0),b=n(3),v=n(18);var g=n(13);function O(e,t){var n=e.indexOf("?"),r=-1!==n?Object(g.parse)(e.substr(n+1)):{};return(-1!==n?e.substr(0,n):e)+"?"+Object(g.stringify)(function(o){for(var e=1;e<arguments.length;e++){var i=null!=arguments[e]?arguments[e]:{},t=Object.keys(i);"function"==typeof Object.getOwnPropertySymbols&&(t=t.concat(Object.getOwnPropertySymbols(i).filter(function(e){return Object.getOwnPropertyDescriptor(i,e).enumerable}))),t.forEach(function(e){var t,n,r;t=o,r=i[n=e],n in t?Object.defineProperty(t,n,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[n]=r})}return o}({},r,t))}n(25);var w,j,_=function(e){function t(){return s()(this,t),l()(this,d()(t).apply(this,arguments))}return m()(t,e),a()(t,[{key:"render",value:function(){var e=this.props,t=e.users,n=e.linkTo,r=e.showDescription;if(e.isLoading)return Object(y.createElement)(b.Spinner,null);if(!t)return Object(y.createElement)("p",null,Object(f.__)("There are no users to show.","wp-team-list"));var o=t.map(function(e){return Object(y.createElement)(S,{user:e,showDescription:r,key:e.id})});return Object(y.createElement)(y.Fragment,null,o,n&&Object(y.createElement)("a",{href:n,className:"show-all"},Object(f.__)("Show all team members","wp-team-list")))}}]),t}(y.Component),x=Object(o.withSelect)(function(e,t){var n=t.number,r=t.roles,o=t.orderBy,i=t.order,s=t.showLink,u=t.showDescription,a=e("wp-team-list/users"),c=a.getUsers,l=a.isLoading,p=O("",{order:i,order_by:o,roles:r,per_page:n,lang:(0,e("core/editor").getEditedPostAttribute)("lang")});return{users:c(p),isLoading:l(p),showLink:s,showDescription:u}})(_),S=function(e){function t(){return s()(this,t),l()(this,d()(t).apply(this,arguments))}return m()(t,e),a()(t,[{key:"render",value:function(){var e=this.props,t=e.user,n=t.id,r=t.role,o=t.role_display_name,i=t.display_name,s=t.avatar_urls,u=t.description,a=t.link,c=t.post_count,l=e.showDescription,p=e.className;return Object(y.createElement)("div",{className:"wp-team-member wp-team-list-item author-".concat(n," role-").concat(r," ").concat(p)},Object(y.createElement)("figure",{className:"wp-team-member-avatar author-image"},Object(y.createElement)("img",{src:s[48],srcSet:"".concat(s[96]," 2x"),alt:"",width:"90"})),Object(y.createElement)("h2",{className:"wp-team-member-name"},i),o&&Object(y.createElement)("p",{className:"wp-team-member-role"},o),l&&Object(y.createElement)("p",{className:"wp-team-member-description"},u),c&&Object(y.createElement)("p",{className:"wp-team-member-posts-link"},Object(y.createElement)("a",{href:a,title:Object(f.sprintf)(Object(f.__)("View all posts by %s","wp-team-list"),i)},Object(f.sprintf)(Object(f._n)("View %s post","View %s posts",c,"wp-team-list"),c))))}}]),t}(y.Component),E=n(12),P=n.n(E),C=n(10),k=n.n(C),A=n(19),I=n.n(A),N=n(20),L=(n(28),function(e){function t(){var e;return s()(this,t),(e=l()(this,d()(t).apply(this,arguments))).setPost=e.setPost.bind(P()(P()(e))),e.suggestPost=e.suggestPost.bind(P()(P()(e))),e}return m()(t,e),a()(t,[{key:"setPost",value:function(e){e&&(this.post=e,(0,this.props.onUpdatePost)(e))}},{key:"suggestPost",value:function(e,t){if(!(e.length<2)){var n=this.props.searchablePostTypes||["post"],r="?subtype=".concat(n.join(","),"&search=").concat(encodeURIComponent(e));k()({path:"/wp/v2/search/"+r}).then(function(e){t(e)})}}},{key:"render",value:function(){var e=this.props,t=e.post,n=e.instanceId,r=e.label,o=e.help,i=e.placeholder,s="post-selector-"+n,u=this.post||t;return Object(y.createElement)(b.BaseControl,{id:s,label:r,help:o},Object(y.createElement)(I.a,{id:s,minLength:2,showAllValues:!0,defaultValue:u?u.title.rendered:"",autoselect:!0,displayMenu:"overlay",onConfirm:this.setPost,source:Object(N.debounce)(this.suggestPost,300),showNoResultsFound:!0,placeholder:i,tStatusQueryTooShort:function(e){return Object(f.sprintf)(Object(f.__)("Type in %s or more characters for results","schilling-content-types"),e)},tNoResults:function(){return Object(f.__)("No results found","schilling-content-types")},tStatusNoResults:function(){return Object(f.__)("No search results.","schilling-content-types")},tStatusSelectedOption:function(e,t){return Object(f.sprintf)(Object(f.__)("%1$s (1 of %2$s) is selected","schilling-content-types"),e,t)},tStatusResults:function(e,t){return Object(y.createElement)("span",null,Object(f.sprintf)(Object(f._n)("%1$s result is available. %2$s","%s results are available. %2$s",e,"schilling-content-types"),e,t))},templates:{inputValue:function(e){return e?e.title:""},suggestion:function(e){return e?e.title:""}}}))}}]),t}(y.Component)),T=function(e){function t(){return s()(this,t),l()(this,d()(t).apply(this,arguments))}return m()(t,e),a()(t,[{key:"render",value:function(){var e=this.props,t=e.className,n=e.post,r=e.attributes,o=r.number,i=r.showDescription,s=r.roles,u=r.orderBy,a=r.order,c=e.setAttributes;return Object(y.createElement)(y.Fragment,null,Object(y.createElement)(x,{number:o,linkTo:n&&n.link,showDescription:i,roles:s,orderBy:u,order:a,className:t}),Object(y.createElement)(v.InspectorControls,null,Object(y.createElement)(b.PanelBody,null,Object(y.createElement)(b.RangeControl,{label:Object(f.__)("Number of users to display","wp-team-list"),value:o,onChange:function(e){return c({number:e})},min:1,max:100}),Object(y.createElement)(b.SelectControl,{label:Object(f.__)("Roles","wp-team-list"),help:Object(f.__)("Only show users with the selected roles","wp-team-list"),value:s,options:[{value:"administrator",label:"Administrator"},{value:"subscriber",label:"Subscriber"},{value:"editor",label:"Editor"}],multiple:!0,onChange:function(e){c({roles:e})}}),Object(y.createElement)(b.SelectControl,{label:Object(f.__)("Order By","wp-team-list"),value:u,options:[{value:"post_count",label:Object(f.__)("Post Count","wp-team-list")},{value:"first_name",label:Object(f.__)("First Name","wp-team-list")},{value:"last_name",label:Object(f.__)("Last Name","wp-team-list")}],onChange:function(e){c({orderBy:e})}}),Object(y.createElement)(b.SelectControl,{label:Object(f.__)("Order","wp-team-list"),value:a,options:[{value:"asc",label:Object(f.__)("Ascending","wp-team-list")},{value:"desc",label:Object(f.__)("Descending","wp-team-list")}],onChange:function(e){c({order:e})}}),Object(y.createElement)(b.ToggleControl,{label:Object(f.__)("Show user description","wp-team-list"),checked:i,onChange:function(e){c({showDescription:e})}}),Object(y.createElement)(L,{label:Object(f.__)("Link to","wp-team-list"),help:Object(f.__)("Select a team page to link to.","wp-team-list"),searchablePostTypes:["page"],post:n,onUpdatePost:function(e){var t=e.id,n=e.subtype;c({postId:t,postType:n})}}),n&&Object(y.createElement)("p",null,Object(y.createElement)(b.Button,{onClick:function(){c({postId:null,postType:null})},isLink:!0,isDestructive:!0},Object(f.__)("Clear selection","wp-team-list"))))))}}]),t}(y.Component),D=Object(o.withSelect)(function(e,t){var n=t.attributes,r=n.postId,o=n.postType;return{post:(0,e("wp-team-list/posts").getPost)(r,o)}})(T),R={title:Object(f.__)("Team List","wp-team-list"),description:Object(f.__)("Display website authors.","wp-team-list"),category:"widgets",icon:"groups",keywords:[Object(f.__)("team","wp-team-list"),Object(f.__)("list","wp-team-list"),Object(f.__)("authors","wp-team-list")],supports:{anchor:!1,customClassName:!1,html:!1},attributes:{number:{type:"number",default:10},postId:{type:"number"},postType:{type:"string"},showDescription:{type:"boolean",default:!0},roles:{type:"array",default:["administrator"]},orderBy:{type:"string",default:"post_count"},order:{type:"string",default:"desc"}},edit:D,save:function(){return null}},F=n(14),B=n.n(F),M=n(11),q=n.n(M),V=n(9),U=n.n(V),Q={users:{},isLoading:{}},H={reducer:function(){var e=0<arguments.length&&void 0!==arguments[0]?arguments[0]:Q,t=1<arguments.length?arguments[1]:void 0;switch(t.type){case"SET_USERS":return U()({},e,{users:U()({},e.users,q()({},t.query,t.users))});case"SET_IS_LOADING":return U()({},e,{isLoading:U()({},e.isLoading,q()({},t.query,t.isLoading))})}return e},actions:{setUsers:function(e,t){return{type:"SET_USERS",query:e,users:t}},setLoading:function(e,t){return{type:"SET_IS_LOADING",query:e,isLoading:t}}},selectors:{getUsers:function(e,t){return e.users[t]},isLoading:function(e,t){var n=e.isLoading;return n[t]}},resolvers:{getUsers:(w=B()(regeneratorRuntime.mark(function e(t){var n;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return Object(o.dispatch)("wp-team-list/users").setLoading(t,!0),e.prev=1,e.next=4,k()({path:"/wp-team-list/v1/users?"+t});case 4:n=e.sent,Object(o.dispatch)("wp-team-list/users").setUsers(t,n),e.next=11;break;case 8:e.prev=8,e.t0=e.catch(1),console.log("[WP Team List] Error ".concat(e.t0.code," while trying to load users: ").concat(e.t0.message));case 11:return e.prev=11,Object(o.dispatch)("wp-team-list/users").setLoading(t,!1),e.finish(11);case 14:case"end":return e.stop()}},e,this,[[1,8,11,14]])})),function(e){return w.apply(this,arguments)})}},$={posts:{}},K={reducer:function(){var e=0<arguments.length&&void 0!==arguments[0]?arguments[0]:$,t=1<arguments.length?arguments[1]:void 0;switch(t.type){case"SET_POST":return U()({},e,{posts:U()({},e.posts,q()({},t.postId,t.post))})}return e},actions:{setPost:function(e,t){return{type:"SET_POST",postId:e,post:t}}},selectors:{getPost:function(e,t){return e.posts[t]||null}},resolvers:{getPost:(j=B()(regeneratorRuntime.mark(function e(t,n){var r;return regeneratorRuntime.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:if(t)return e.next=3,k()({path:"/wp/v2/".concat(n,"s/").concat(t,"?_embed")});e.next=6;break;case 3:e.t0=e.sent,e.next=7;break;case 6:e.t0=null;case 7:r=e.t0,Object(o.dispatch)("wp-team-list/posts").setPost(t,r);case 9:case"end":return e.stop()}},e,this)})),function(e,t){return j.apply(this,arguments)})}};Object(o.registerStore)("wp-team-list/users",H),Object(o.registerStore)("wp-team-list/posts",K),Object(r.registerBlockType)("required/wp-team-list",R)}]);
