+++ assets/js/src/blocks/index.js
@@ -1,0 +1,1 @@
+export { name as TeamListName, settings as TeamListSettings } from './team-list';
