+++ assets/js/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as TeamList } from './team-list';
+export { default as TeamMember } from './team-member';
+export { default as PostSelector } from './post-selector';
