+++ assets/js/src/components/team-list/editor.css
