+++ package.json
@@ -27,25 +27,29 @@
     "watch:js": "webpack --watch"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
-    "@babel/preset-env": "^7.1.0",
-    "@wordpress/api-fetch": "^2.2.3",
+    "@babel/cli": "^7.2.0",
+    "@babel/core": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
+    "@wordpress/api-fetch": "^2.2.5",
     "@wordpress/babel-plugin-import-jsx-pragma": "^1.1.2",
     "@wordpress/babel-preset-default": "^3.0.1",
-    "@wordpress/components": "^6.0.1",
-    "@wordpress/data": "^3.1.2",
-    "@wordpress/i18n": "^3.0.1",
-    "autoprefixer": "^9.3.1",
+    "@wordpress/blocks": "^6.0.3",
+    "@wordpress/components": "^7.0.3",
+    "@wordpress/data": "^4.0.1",
+    "@wordpress/editor": "^9.0.4",
+    "@wordpress/element": "^2.1.8",
+    "@wordpress/i18n": "^3.1.0",
+    "@wordpress/url": "^2.3.1",
+    "autoprefixer": "^9.4.1",
     "babel-core": "^6.26.3",
     "babel-loader": "^8.0.4",
-    "concurrently": "^4.0.1",
+    "concurrently": "^4.1.0",
     "css-loader": "^1.0.1",
     "css-mqpacker": "^7.0.0",
     "cssnano": "^4.1.7",
     "eslint-config-wordpress": "^2.0.0",
-    "mini-css-extract-plugin": "^0.4.3",
-    "postcss": "^7.0.5",
+    "mini-css-extract-plugin": "^0.4.5",
+    "postcss": "^7.0.6",
     "postcss-cli": "^6.0.0",
     "postcss-hexrgba": "^1.0.0",
     "postcss-import": "^12.0.0",
@@ -56,7 +60,10 @@
     "raw-loader": "^0.5.1",
     "rtlcss": "^2.4.0",
     "uglifyjs-webpack-plugin": "^2.0.1",
-    "webpack": "^4.25.1",
+    "webpack": "^4.27.0",
     "webpack-cli": "^3.1.2"
+  },
+  "dependencies": {
+    "accessible-autocomplete": "^1.6.2"
   }
 }
